CIS 278 C++ Lab 1
made a loop which iterates 10 times. Each iteration it will
generate a random number between 0-3 using the modulo operation.
Each numer, 0-3 will correspond to an operation which will then be chosen to
produce a question.
Each question will have 2 numbers, each generated randomly and between values 1-100
inclusive. This is done intentionally to avoid divisionByZero errors.
The program then takes the user input for the question and compares it to the
actual integer value of the solution. If correct, the user is notified
and their score is increased. If incorrect, the user is notified of the correct
solution and their score does not increase.
Additionally, I added if the user wanted to play again or exit the program using a
simple boolean statment and a while loop.
The program assumes the user will enter the correct type for each input, no
exceptions or errors taken into account